+++ kidsability/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styling/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
